+++ Scribe/Scribe-Core/src/utopia/scribe/core/model/cached/logging/StackTrace.scala
@@ -30,9 +30,10 @@ object StackTrace extends FromModelFactory[StackTrace]
 	private val methodSplitter = Regex.escape('$')
 	
 	private lazy val schema = ModelDeclaration(
-		PropertyDeclaration("class", StringType, Vector("className", "class_name"), "CouldNotParse"),
+		PropertyDeclaration("file", StringType, Vector("fileName", "file_name"), "CouldNotParse"),
+		PropertyDeclaration("class", StringType, Vector("className", "class_name"), isOptional = true),
 		PropertyDeclaration("method", StringType, Vector("methodName", "method_name", "function"), "couldNotParse"),
-		PropertyDeclaration("line", IntType, Vector("lineNumber", "line_number"), -1),
+		PropertyDeclaration("line", IntType, Vector("lineNumber", "line_number"), isOptional = true),
 		PropertyDeclaration("cause", ModelType, isOptional = true)
 	)
 	
@@ -41,7 +42,8 @@ object StackTrace extends FromModelFactory[StackTrace]
 	
 	override def apply(model: ModelLike[Property]): Try[StackTrace] =
 		schema.validate(model).toTry.map { model =>
-			apply(model("class"), model("method"), model("line"), model("cause").model.flatMap { apply(_).toOption })
+			apply(model("file"), model("class"), model("method"), model("line"),
+				model("cause").model.flatMap { apply(_).toOption })
 		}
 	
 	
@@ -66,6 +68,7 @@ object StackTrace extends FromModelFactory[StackTrace]
 	}
 	// nextElement will be called until it returns None
 	private def _from(element: JStackTraceElement)(nextElement: => Option[JStackTraceElement]): StackTrace = {
+		val fileName = Option(element.getFileName).map { _.untilLast(".") }
 		// Converts the dollar-sign -ending class names to those that don't end with a dollar sign
 		// Removes the package part, also
 		val className = NotEmpty(element.getClassName.split('.').takeRightWhile { _.head.isUpper }.toVector) match {
@@ -78,7 +81,16 @@ object StackTrace extends FromModelFactory[StackTrace]
 			// Also ignores empty names (e.g. "flatMap$" now becomes "flatMap" instead of "")
 			.reverseIterator.filterNot { _.forall { _.isDigit } }
 			.nextOption().getOrElse("")
-		apply(className, methodName, element.getLineNumber, nextElement.map { _from(_)(nextElement) })
+		
+		// Doesn't include className if it is duplicate with the file name.
+		// Also makes sure some file name is specified.
+		val (appliedFileName, appliedClassName) = fileName match {
+			case Some(fName) => (fName, if (className == fName) "" else className)
+			case None => (className.nonEmptyOrElse("Unknown"), "")
+		}
+		
+		apply(appliedFileName, appliedClassName, methodName, Option(element.getLineNumber).filter { _ >= 0 },
+			nextElement.map { _from(_)(nextElement) })
 	}
 }
 
@@ -88,35 +100,50 @@ object StackTrace extends FromModelFactory[StackTrace]
   * @author Mikko Hilpinen
   * @since 23.5.2023, v0.1
   * @constructor Creates a new stack trace element
-  * @param className Name of the class where this event occurred
+  * @param fileName Name of the file in which the targeted class appears
+  * @param className Name of the class where this event occurred. Empty if identical to the file name.
   * @param methodName Name of the method where this event occurred
   * @param lineNumber Index of the line (1-based) where this event occurred
   * @param cause Stack trace element / event that occurred before this element. None if this was the first occurrence.
   */
-// TODO: LineNumber should be optional
-case class StackTrace(className: String, methodName: String, lineNumber: Int, cause: Option[StackTrace] = None)
+case class StackTrace(fileName: String, className: String, methodName: String, lineNumber: Option[Int] = None,
+                      cause: Option[StackTrace] = None)
 	extends ModelConvertible
 {
 	// COMPUTED -------------------------
 	
 	/**
+	  * @return Name of the file and class in which this trace appears.
+	  *         If the two are identical, only returns the file name.
+	  */
+	def fileAndClassName = if (className.isEmpty) fileName else s"$fileName: $className"
+	
+	/**
 	  * @return A string that represents this individual stack trace element
 	  */
-	def logLine = s"$className.$methodName (line $lineNumber)"
+	def logLine = {
+		val linePart = lineNumber match {
+			case Some(n) => s" (line $n)"
+			case None => ""
+		}
+		s"$fileAndClassName.$methodName$linePart"
+	}
 	/**
 	  * @return An iterator that returns one line for each stack trace element in this stack
 	  */
 	def logLinesIterator = {
-		// Groups by class, then by method (consecutive only)
-		topToBottomIterator.groupBy { _.className }.flatMap { case (className, elements) =>
-			val prefix = MutatingOnce(className)("\t")
-			elements.iterator.groupBy { _.methodName }.map { case (methodName, elements) =>
-				// Groups multiple line number instances to a single set
-				val lineNumberStr = elements.oneOrMany match {
-					case Left(element) => element.lineNumber.toString
-					case Right(elements) => s"[${elements.map { _.lineNumber }.mkString(", ")}]"
+		// Groups by file, class and method (consecutive only)
+		topToBottomIterator.groupBy { _.fileName }.flatMap { case (_, elements) =>
+			elements.iterator.groupBy { _.fileAndClassName }.flatMap { case (className, elements) =>
+				val prefix = MutatingOnce(className)("\t")
+				elements.iterator.groupBy { _.methodName }.map { case (methodName, elements) =>
+					// Groups multiple line number instances to a single set
+					val lineNumberStr = elements.flatMap { _.lineNumber }.oneOrMany match {
+						case Left(number) => s": $number"
+						case Right(numbers) => if (numbers.isEmpty) "" else s": [${ numbers.mkString(", ") }]"
+					}
+					s"${ prefix.value }.$methodName$lineNumberStr"
 				}
-				s"${prefix.value}.$methodName: $lineNumberStr"
 			}
 		}
 	}
@@ -143,6 +170,6 @@ case class StackTrace(className: String, methodName: String, lineNumber: Int, ca
 	
 	override def toString = logLinesIterator.mkString("\n")
 	
-	override def toModel: Model =
-		Model.from("class" -> className, "method" -> methodName, "line" -> lineNumber, "cause" -> cause)
+	override def toModel: Model = Model.from("file" -> fileName, "class" -> className,
+		"method" -> methodName, "line" -> lineNumber, "cause" -> cause)
 }